fix the index append (#156)
the index append introduced in #155 created a valid index but didn't
properly append the image contents to it. instead it was simply
appending an index manifest to the index 🤦

this PR fixes that and adds another test to catch it (I think?). ideally
I'd like the test to push up a platform specific image(s) but I can't
figure out how to do that with `random.Image`.

I've at least tested this locally and it appears to be generating a
proper index:

```
{
  "schemaVersion": 2,
  "mediaType": "application/vnd.oci.image.index.v1+json",
  "manifests": [
    {
      "mediaType": "application/vnd.oci.image.manifest.v1+json",
      "size": 562,
      "digest": "sha256:c5316b34c4564257960ce18e73b6f421cae0e67c5e3cf18a97cd755947c071c4",
      "platform": {
        "architecture": "amd64",
        "os": "linux"
      }
    },
    {
      "mediaType": "application/vnd.oci.image.manifest.v1+json",
      "size": 561,
      "digest": "sha256:dad0b65b0a055bcee31b5888d5435c5f73e25785f0e92be473996fb060bb62eb",
      "platform": {
        "architecture": "arm64",
        "os": "linux"
      }
    }
  ]
}
```