Use tarfs to implement structure test (#228)
Previously, we scanned through the image filesystem and matched the
tar.Header.Name field against the file path given by the test. This
works most of the time, but doesn't handle symlinks at all.

This change writes the entire layer tar to a temporary file, parses it
with tarfs to get an fs.FS, attempts to randomly access each file from
the test, and deletes the temporary file afterwards.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>