Add structure test support for optional files, directories, and permissions (#253)
Currently, structure tests allow for specific file permission checks;
while this works for files that are guaranteed to exist, there are other
files and paths that may not exist universally but may also be too
numerous to check for on a per-module caller basis.

This PR addresses this gap by allowing for files to be classified as
optional. If a structure test encounters an optional file that does not
exist, that file will be ignored instead of throwing an error (this
behavior is opt-in, as well).

There are also cases where we care about more than one file in a
directory (and maybe a specific directory?). To address this, the PR
adds support for top-level directory permission checks as well recursive
checks to ensure that every path within a given directory has the
expected permission. For the latter, there is also an opt-in input to
limit the check to files within the given directory.

There's also a possibility that we want to fail a test if any paths with
a certain permission are discovered which is supported via a new
`permissions` block. Overrides can also be specified for this check if
there are documented exceptions.

---------

Signed-off-by: egibs <20933572+egibs@users.noreply.github.com>